- Added more boundary checks to the FD_#?() macros which now also catch   the case of negative socket IDs.

+++ include/sys/select.h
@@ -72,9 +72,9 @@ typedef	struct fd_set
 	unsigned long bits[(FD_SETSIZE + 31) / 32];
 } fd_set;
 
-#define	FD_SET(n,p)		((void)((n) < FD_SETSIZE ? (p)->bits[((unsigned long)n) >> 5] |=  (1UL << (((unsigned long)n) & 31)) : 0))
-#define	FD_CLR(n,p)		((void)((n) < FD_SETSIZE ? (p)->bits[((unsigned long)n) >> 5] &= ~(1UL << (((unsigned long)n) & 31)) : 0))
-#define	FD_ISSET(n,p)	((n) < FD_SETSIZE && ((p)->bits[((unsigned long)n) >> 5] & (1UL << (((unsigned long)n) & 31))) != 0)
+#define	FD_SET(n,p)		((void)(((unsigned long)n) < FD_SETSIZE ? (p)->bits[((unsigned long)n) >> 5] |=  (1UL << (((unsigned long)n) & 31)) : 0))
+#define	FD_CLR(n,p)		((void)(((unsigned long)n) < FD_SETSIZE ? (p)->bits[((unsigned long)n) >> 5] &= ~(1UL << (((unsigned long)n) & 31)) : 0))
+#define	FD_ISSET(n,p)	(((unsigned long)n) < FD_SETSIZE && ((p)->bits[((unsigned long)n) >> 5] & (1UL << (((unsigned long)n) & 31))) != 0)
 #define	FD_COPY(f,t)	((void)memmove(t,f,sizeof(*(f))))
 #define	FD_ZERO(p)		((void)memset(p,0,sizeof(*(p))))
